+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Integration test for cluster_controller written with envtest and ginkgo
+
 ## [0.0.8] - 2020-12-16
 
 ### Added